fix(viz-c): three-round review fixes — rounding parity, host lifetimes, build hygiene
Review A (Python semantics): same_time keeps np.isclose's rtol term;
fence/probe/rescale sampling rounds half-to-even (nearbyint) before
clamping like Python round(); colorize_amplitude quantizes the P98 limit
to float32 exactly like the numpy path; clamp_indices fails closed on
non-finite input (Python round raises); set_orthogonal_slice_indices
keeps an absent index absent (reconcile seeds the middle line).

Review B (lifecycle/concurrency): open_volume's finished callback uses
the pointer any_cast form (a failed outcome carries an empty any — the
value form would throw on the GUI thread); time_map_/volume_owner_ are
QPointers and the host destructor unbinds the map's scene pointer
(cross-parent dangling closed both directions); late callbacks only
clear the owner they belong to; X11 is a soft probe (broken-GLX guard
degrades out, never a configure failure); threading comments state the
real contract (GUI thread is the volume's single reader); fault-hook and
snapshot_cache_ thread boundaries documented.

Review C (product/build): accidentally committed a.out removed +
.gitignored; example destruction order fixed (JobCenter dies before the
host, mirroring the MainWindow member ordering) — offscreen and
software-GL runs now exit 0 where the pre-fix process died with a
JobOwner dual-ownership UAF at teardown; dock wires restore_state() so
the persisted joint state actually comes back; missing-fixture example
degradation is announced.

+++ libs/geo3d_viz/src/joint/color_scales.cpp
@@ -88,8 +88,13 @@ std::vector<std::uint8_t> colorize_amplitude(
     for (float v : amplitude) {
         if (std::isfinite(v)) finite.push_back(std::fabs(static_cast<double>(v)));
     }
+    // Python computes the percentile over a float32 array, so the limit
+    // itself carries float32 quantization; reproduce it exactly.
     double limit =
-        finite.empty() ? 1.0 : percentile_linear(finite, 98.0);
+        finite.empty()
+            ? 1.0
+            : static_cast<double>(static_cast<float>(
+                  percentile_linear(finite, 98.0)));
     if (!std::isfinite(limit) || limit <= 1e-12) limit = 1.0;
 
     const std::vector<RgbStop>& stops = seismic_color_scale(color_scale);
